마이페이지 프로필 깜빡거림 수정
