+++ client/src/pages/chat/chatDetail/ChatDetail-1.js
@@ -1,131 +1,140 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-// import axios from "axios";
+import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import defaultAvatar from "../../../assets/img/avatar/avatar.jpg";
 import { ReactComponent as Xmark } from "../../../assets/img/icon/xmark-solid.svg";
-// import { useParams } from "react-router-dom";
-import * as StompJs from "@stomp/stompjs";
-import * as SockJS from "sockjs-client";
 import { useParams } from "react-router-dom";
+import { over } from "stompjs";
+import SockJS from "sockjs-client";
+
+let client = null;
 
 const ChatDetail = () => {
-  const { id } = useParams();
-  const navigate = useNavigate();
-  const ROOM_SEQ = id;
-  const client = useRef({});
-  const [chatMessages, setChatMessages] = useState([]);
-  const [message, setMessage] = useState("");
   const memberId = localStorage.getItem("memberid");
+  const { id } = useParams(); // 룸아이디
+  const navigate = useNavigate();
+
+  const headers = {
+    "Content-Type": "application/json",
+    Authorization: `${localStorage.getItem("authorization")}`,
+  };
+  const [sellerId, setSellerId] = useState("");
+  const [sellerName, setSellerName] = useState("");
+  const [msg, setMsg] = useState(""); // 보내는 메세지
+  const [data, setData] = useState(""); // 받는 메세지
+
+  // const connection = () => {
+  //   let sock = new SockJS(
+  //     `${process.env.REACT_APP_API_URL}/gs-guide-websocket`
+  //   );
+  //   const client = over(sock);
+
+  //   client.connect({}, () => {
+
+  //     client.subscribe(
+  //       `/queue/addChatToClient/${id}`, // 룸아이디
+  //       (chat) => {
+  //         const newMessage = JSON.parse(chat.body);
+  //         setData(data.concat(newMessage));
+  //         console.log(chat);
+  //       },
+  //       {}
+  //     );
+  //     // .unsubscribe();
+  //   });
+  //   // 위의 콜백함수는 클라이언트가 서버로부터 메세지(data)를 전달받을때 실행된다.
+  //   // 특정 채팅방에서 나가거나 해당 어플리케이션에서 나갔을 경우 상대방이 보낸 메세지를 읽지 않음으로 표시하기 위해 구독을 끊어준다.
+  //   // 전달하는 메세지 / 전달받는 메세지를 따로 변수에 담아준다. --> 그래야 css 작업할 수 있을듯. 근데 이게 같은 내용 아닌가?
+  //   // return () => client.disconnect();
+  //   // }, [client, memberId]); // <--- ???
+  // };
+
+  // const sendMessage = () => {
+  //   let chatMessage = {
+  //     chatRoomId: id,
+  //     memberId: memberId, // 로그인한 회원
+  //     content: msg,
+  //   };
+
+  //   client.send(`/app/chat/${id}`, {}, JSON.stringify(chatMessage));
+  //   console.log(chatMessage);
+  //   // setUserData({ ...userData, content: "" });
+  // };
+
+  const getSellerId = async () => {
+    await axios
+      .get(`${process.env.REACT_APP_API_URL}/v1/chat/room/${id}`)
+      .then((res) => {
+        setSellerId(res.data.sellerId);
+        // console.log(res.data.sellerId);
+      })
+      .catch((err) => console.log(err));
+  };
+
+  const getSellerName = async () => {
+    await axios
+      .get(`${process.env.REACT_APP_API_URL}/v1/members/${sellerId}`, {
+        headers: headers,
+      })
+      .then((res) => {
+        setSellerName(res.data.memberName);
+        // console.log(res.data.memberName);
+      })
+      .catch((err) => console.log(err));
+  };
+
+  const handleInput = (e) => {
+    setMsg(e.target.value);
+  };
 
   useEffect(() => {
-    connect();
-
-    return () => disconnect();
+    getSellerId();
+    getSellerName();
+    // connection();
   }, []);
 
-  const connect = () => {
-    client.current = new StompJs.Client({
-      webSocketFactory: () =>
-        new SockJS(`${process.env.REACT_APP_API_URL}/gs-guide-websocket`), // proxy를 통한 접속
-      debug: function (str) {
-        console.log(str);
-      },
-      reconnectDelay: 5000,
-      heartbeatIncoming: 4000,
-      heartbeatOutgoing: 4000,
-      onConnect: () => {
-        subscribe();
-      },
-      onStompError: (frame) => {
-        console.error(frame);
-      },
-    });
-
-    client.current.activate();
-  };
-
-  const disconnect = () => {
-    client.current.deactivate();
-  };
-
-  const subscribe = () => {
-    client.current.subscribe(
-      `/queue/addChatToClient/${ROOM_SEQ}`,
-      ({ body }) => {
-        setChatMessages((_chatMessages) => [
-          ..._chatMessages,
-          JSON.parse(body),
-        ]);
-      }
-    );
-  };
-
-  const publish = (message) => {
-    if (!client.current.connected) {
-      return;
-    }
-
-    client.current.publish({
-      destination: `/app/chat/${ROOM_SEQ}`,
-      body: JSON.stringify({
-        chatRoomId: ROOM_SEQ,
-        memberId: memberId,
-        content: message,
-      }),
-    });
-
-    setMessage("");
-  };
-
   return (
-    <CDContainer>
-      <CDContent>
-        <ContentTop>
-          <AvatarAndNameWrapper>
-            <img src={defaultAvatar} className="avatar" alt="유저이미지"></img>
-            <div className="name">{}</div>
-          </AvatarAndNameWrapper>
-          <XmarkWrapper
-            onClick={() => {
-              navigate(-1);
-            }}
-          >
-            <Xmark />
-          </XmarkWrapper>
-        </ContentTop>
-        <ContentMiddle>
-          {chatMessages && chatMessages.length > 0 && (
-            <ul>
-              {chatMessages.map((el, index) => (
-                <li
-                  key={index}
-                  className={
-                    el.memberId === Number(memberId)
-                      ? "login-user-me"
-                      : "login-user-you"
-                  }
-                >
-                  {el.memberId}: {el.content}
-                </li>
-              ))}
-            </ul>
-          )}
-        </ContentMiddle>
-        <ContentBottom>
-          <textarea
-            type={"text"}
-            placeholder={"message"}
-            value={message}
-            onChange={(e) => setMessage(e.target.value)}
-            onKeyPress={(e) => e.which === 13 && publish(message)}
-          />
-          <button onClick={() => publish(message)}>send</button>
-        </ContentBottom>
-      </CDContent>
-    </CDContainer>
+    <>
+      <CDContainer>
+        <h1>채팅</h1>
+        <div>{data}</div>
+        <CDContent>
+          <ContentTop>
+            <AvatarAndNameWrapper>
+              <img src={defaultAvatar} className="avatar"></img>
+              <div className="name">{sellerName}</div>
+            </AvatarAndNameWrapper>
+            <XmarkWrapper
+              onClick={() => {
+                navigate(`/`);
+              }}
+            >
+              <Xmark />
+            </XmarkWrapper>
+          </ContentTop>
+          <ContentMiddle></ContentMiddle>
+          <ContentBottom>
+            <textarea
+              value={msg}
+              onChange={handleInput}
+              placeholder="메세지를 입력하세요"
+            />
+            <button
+            // onClick={() => {
+            //   sendMessage(msg);
+            //   setMsg("");
+            // }}
+            >
+              전송
+            </button>
+          </ContentBottom>
+        </CDContent>
+      </CDContainer>
+    </>
   );
 };
+
 export default ChatDetail;
 
 const CDContainer = styled.div`
@@ -135,14 +144,12 @@ const CDContainer = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
-  padding: 3.6rem 0 3.1rem 0;
 
   h1 {
     text-align: center;
     margin: 3.8rem auto;
     font-size: 2.5rem;
     font-family: "NotoSansKR-Medium";
-
     @media ${(props) => props.theme.mobile} {
       margin: 3.6rem 0 3.1rem 0;
       font-size: 1.5rem;
@@ -154,10 +161,9 @@ const CDContent = styled.div`
   display: flex;
   flex-direction: column;
   margin: 0 auto 8.875rem auto;
-
   @media ${(props) => props.theme.mobile} {
-    width: 22.5625rem;
-    height: 33.5625rem;
+    /* width: 22.5625rem;
+    height: 33.5625rem; */
     margin: 0 auto 8.875rem auto;
   }
 `;
@@ -172,7 +178,6 @@ const ContentTop = styled.div`
   border: 0.31rem solid ${(props) => props.theme.gray4};
   border-radius: 1.25rem;
   background-color: ${(props) => props.theme.bgColor};
-
   @media ${(props) => props.theme.mobile} {
     width: 22.25rem;
     height: 4.5625rem;
@@ -185,7 +190,6 @@ const AvatarAndNameWrapper = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
-
   .avatar {
     object-fit: contain;
     width: 3rem;
@@ -193,18 +197,15 @@ const AvatarAndNameWrapper = styled.div`
     border-radius: 50%;
     background-color: ${(props) => props.theme.gray4};
     margin-left: 1rem;
-
     @media ${(props) => props.theme.mobile} {
       width: 2.625rem;
       height: 2.625rem;
     }
   }
-
   .name {
     font-size: 1.5rem;
     font-family: "NotoSansKR-Bold";
     margin-left: 1.8rem;
-
     @media ${(props) => props.theme.mobile} {
       font-size: 1.3rem;
       margin-left: 1.3rem;
@@ -216,11 +217,6 @@ const XmarkWrapper = styled.div`
   width: 1.3rem;
   height: 1.3rem;
   margin: -0.7rem 1rem 0 0rem;
-  svg {
-    fill: ${(props) => props.theme.gray4};
-    cursor: pointer;
-  }
-
   @media ${(props) => props.theme.mobile} {
     width: 0.9rem;
     height: 0.9rem;
@@ -229,55 +225,17 @@ const XmarkWrapper = styled.div`
 `;
 
 const ContentMiddle = styled.div`
-  display: flex;
-  flex-direction: column-reverse;
   width: 36.5rem;
   height: 25rem;
   margin-bottom: 1.5rem;
   border: 0.31rem solid ${(props) => props.theme.gray4};
   border-radius: 1.25rem;
   background-color: ${(props) => props.theme.bgColor};
-  overflow: auto;
-  font-family: "NotoSansKR-Medium";
-  font-size: 20px;
-  padding: 1rem;
-
-  li {
-    padding: 4px 0;
-  }
-
-  ::-webkit-scrollbar {
-    width: 1rem;
-  }
-
-  ::-webkit-scrollbar-thumb {
-    background-color: ${(props) => props.theme.gray4};
-    border-radius: 1.25rem;
-  }
-
-  .login-user-me {
-    display: flex;
-    flex-direction: row-reverse;
-    /* background-color: ${(props) => props.theme.primary}; */
-    padding: 8px 8px;
-    margin: 8px 0;
-    border-radius: 15px;
-  }
-
-  .login-user-me {
-    display: flex;
-    /* background-color: ${(props) => props.theme.primary}; */
-    padding: 8px 8px;
-    margin: 8px 0;
-    border-radius: 15px;
-  }
-
   @media ${(props) => props.theme.mobile} {
     width: 22.25rem;
     height: 21.6875rem;
     margin-bottom: 1.375rem;
     border: 0.1875rem solid ${(props) => props.theme.gray4};
-    font-size: 16px;
   }
 `;
 const ContentBottom = styled.div`
@@ -288,7 +246,6 @@ const ContentBottom = styled.div`
   border-radius: 1.25rem;
   background-color: ${(props) => props.theme.bgColor};
   display: flex;
-
   @media ${(props) => props.theme.mobile} {
     width: 22.25rem;
     height: 4.5625rem;
@@ -305,18 +262,15 @@ const ContentBottom = styled.div`
     color: ${(props) => props.theme.textColor};
     font-size: 1.25rem;
     font-family: "NotoSansKR-Bold";
-
     @media ${(props) => props.theme.mobile} {
       width: 15rem;
       height: 2.4rem;
       font-size: 1rem;
     }
-
     :focus {
       outline: none;
     }
   }
-
   button {
     width: 6.375rem;
     height: 2.8125rem;
@@ -326,14 +280,12 @@ const ContentBottom = styled.div`
     margin: 4.78rem 0 0.8rem 0;
     background-color: ${(props) => props.theme.primary};
     color: ${(props) => props.theme.white};
-
     @media ${(props) => props.theme.mobile} {
       width: 4.375rem;
       height: 2.125rem;
       font-size: 1rem;
       margin: 1.49rem 0.4rem 0.8rem 0;
     }
-
     :hover {
       filter: drop-shadow(0rem 0.2rem 0.2rem ${(props) => props.theme.gray4});
     }
